% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sina_plot.R
\name{sina_plot}
\alias{sina_plot}
\title{Make a sina plot of the Shapley values computed using shapr.}
\usage{
sina_plot(explanation)
}
\arguments{
\item{explanation}{shapr list containing an explanation produced by shapr::explain.}
}
\value{
ggplot2 object containing the sina plot.
}
\description{
Make a sina plot of the Shapley values computed using shapr.
}
\details{
Function adapted from \link[SHAPforxgboost]{shap.plot.summary}.
Copyright © 2020 - Yang Liu & Allan Just
}
\examples{
# set parameters and random seed
set.seed(2020)
N <- 1000
m <- 4
sds <- runif(4, 0.5, 1.5)
pars <- runif(7, -1, 1)

# Create data from a structural equation model
X_1 <- rnorm(N, sd = sds[1])
Z <- rnorm(N, 1)
X_2 <- X_1 * pars[1] + Z * pars[2] + rnorm(N, sd = sds[2])
X_3 <- X_1 * pars[3] + Z * pars[4] + rnorm(N, sd = sds[3])
Y <- X_1 * pars[5] + X_2 * pars[6] + X_3 * pars[7] + rnorm(N, sd = sds[4])

# collecting data
mu_A <- rep(0, m)
X_A <- cbind(X_1, X_2, X_3)
dat_A <- cbind(X_A, Y)
cov_A <- cov(dat_A)

model <- lm(Y ~ . + 0 , data = as.data.frame(dat_A))
explainer <- shapr::shapr(X_A, model)
y_mean <- mean(Y)

explanation_classic <- shapr::explain(
  dat_A, 
  approach = "gaussian", 
  explainer = explainer, 
  prediction_zero = y_mean
)
sina_plot(explanation_classic)

explanation_causal <- shapr::explain(
  dat_A, 
  approach = "causal", 
  explainer = explainer, 
  prediction_zero = y_mean, 
  ordering = list(1, c(2, 3))
)
sina_plot(explanation_causal)

}
\seealso{
\link[SHAPforxgboost]{shap.plot.summary}
}
